requirements.txt
fastapi==0.111.0
PyJWT==2.8.0
python-dotenv==1.0.1
tensorflow==2.16.1
tensorflow-io==0.31.0
google-cloud-storage==2.17.0
numpy==1.26.4
pandas==2.2.2
